backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
aioredis==2.0.1
celery==5.3.4

# Audio/Video Processing
openai-whisper==20250625
ffmpeg-python==0.2.0
yt_dlp==2025.8.20
pydub==0.25.1
librosa==0.11.0
soundfile==0.13.1
scipy==1.16.1
noisereduce==3.0.3

# PyTorch with CUDA support
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Speaker diarization
pyannote.audio[all]>=3.1.0
pyannote.core>=5.0.0

# Optimization
numpy>=2.0.0
numba>=0.61.0
tiktoken>=0.11.0

# AI/ML Libraries
openai>=1.0.0
groq>=0.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.5.1
faiss-cpu>=1.7.4

# Utils
requests>=2.26.0
tqdm>=4.67.0
validators>=0.22.0